requirements.txt
pytest>=6.2.2
importlib-resources>=5.1.0
flake8>=7.0.0
pandas>=2.1.4
pytest>=6.2.2
importlib-resources>=5.1.0
cached-property>=1.5.2
scipy>=1.12.0
biopython>=1.83